Clean up code structure, clean up warnings
