--- src/state.rs
@@ -1,4 +1,0 @@
-pub use self::state::{};
-
-mod state {
-}
